Add Interpolant, CenterMagInterpolant
Each parameter needs knowledge of how the values are to be interpolated.
The colors parameter, for instance, isn't just a simple interpolation of
floating-point values, but needs to understand color cycling as well as
color interpolation.  Encapsulate that knowledge behind the Interpolant
interface and use a factory function to create the appropriate
Interpolant implementation for the parameter.